+++ src/main/java/com/github/ddth/kafka/internal/KafkaHelper.java
@@ -1,11 +1,13 @@
 package com.github.ddth.kafka.internal;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Properties;
 import java.util.Set;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.clients.producer.KafkaProducer;
@@ -81,31 +83,46 @@ public class KafkaHelper {
      */
     public static void seekToEnd(final KafkaConsumer<?, ?> consumer, final String topic) {
         synchronized (consumer) {
-            // first, save the current subscription
-            Set<String> subscription = consumer.subscription();
-            try {
-                // second, unsubscribe and re-subscribe to all partitions.
-                consumer.unsubscribe();
-                List<PartitionInfo> partInfo = consumer.partitionsFor(topic);
-                Collection<TopicPartition> tpList = new ArrayList<>();
-                for (PartitionInfo p : partInfo) {
-                    TopicPartition tp = new TopicPartition(topic, p.partition());
-                    tpList.add(tp);
-                }
-                consumer.assign(tpList);
-                consumer.seekToEnd(tpList);
-                // we want to seek as soon as possible
-                for (TopicPartition tp : tpList) {
-                    // since seekToEnd evaluates lazily, invoke position() so
-                    // that seeking will be committed.
-                    consumer.position(tp);
+            Set<TopicPartition> topicParts = consumer.assignment();
+            if (topicParts != null) {
+                for (TopicPartition tp : topicParts) {
+                    if (StringUtils.equals(topic, tp.topic())) {
+                        consumer.seekToEnd(Arrays.asList(tp));
+                        // we want to seek as soon as possible
+                        // since seekToEnd evaluates lazily, invoke position()
+                        // so
+                        // that seeking will be committed.
+                        consumer.position(tp);
+                    }
                 }
                 consumer.commitSync();
-            } finally {
-                // finally, restore subscription
-                consumer.unsubscribe();
-                consumer.subscribe(subscription);
             }
+
+            // // first, save the current subscription
+            // Set<String> subscription = consumer.subscription();
+            // try {
+            // // second, unsubscribe and re-subscribe to all partitions.
+            // consumer.unsubscribe();
+            // List<PartitionInfo> partInfo = consumer.partitionsFor(topic);
+            // Collection<TopicPartition> tpList = new ArrayList<>();
+            // for (PartitionInfo p : partInfo) {
+            // TopicPartition tp = new TopicPartition(topic, p.partition());
+            // tpList.add(tp);
+            // }
+            // consumer.assign(tpList);
+            // consumer.seekToEnd(tpList);
+            // // we want to seek as soon as possible
+            // for (TopicPartition tp : tpList) {
+            // // since seekToEnd evaluates lazily, invoke position() so
+            // // that seeking will be committed.
+            // consumer.position(tp);
+            // }
+            // consumer.commitSync();
+            // } finally {
+            // // finally, restore subscription
+            // consumer.unsubscribe();
+            // consumer.subscribe(subscription);
+            // }
         }
     }
 
@@ -145,23 +162,28 @@ public class KafkaHelper {
         props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG,
                 ByteArraySerializer.class.getName());
 
-        // max request size: 128kb
-        props.put(ProducerConfig.MAX_REQUEST_SIZE_CONFIG, 128L * 1024L);
+        // max request size: 128kb, it's also max message size
+        props.put(ProducerConfig.MAX_REQUEST_SIZE_CONFIG, String.valueOf(128 * 1024));
 
-        // 1mb buffer & 128-record batch size
-        props.put(ProducerConfig.BUFFER_MEMORY_CONFIG, 1L * 1024L * 1024L);
-        props.put(ProducerConfig.BATCH_SIZE_CONFIG, 128);
+        // 60mb buffer & 256-record batch size
+        props.put(ProducerConfig.BUFFER_MEMORY_CONFIG, String.valueOf(60 * 1024 * 1024));
+        props.put(ProducerConfig.BATCH_SIZE_CONFIG, String.valueOf(256));
 
-        props.put(ProducerConfig.LINGER_MS_CONFIG, 1);
+        // KafkaProducer.send() / KafkaProducer.partitionsFor() max block: 60s
+        props.put(ProducerConfig.MAX_BLOCK_MS_CONFIG, String.valueOf(60000));
 
-        // ack timeout 10 seconds
-        props.put(ProducerConfig.REQUEST_TIMEOUT_MS_CONFIG, 10000);
+        // props.put(ProducerConfig.COMPRESSION_TYPE_CONFIG, "snappy");
+
+        props.put(ProducerConfig.LINGER_MS_CONFIG, String.valueOf(10));
+
+        // ack timeout 30 seconds
+        props.put(ProducerConfig.REQUEST_TIMEOUT_MS_CONFIG, String.valueOf(30000));
 
         // metadata fetch timeout: 10 seconds
-        props.put(ProducerConfig.MAX_BLOCK_MS_CONFIG, 10000);
+        props.put(ProducerConfig.MAX_BLOCK_MS_CONFIG, String.valueOf(10000));
 
-        props.put(ProducerConfig.RETRIES_CONFIG, 3);
-        props.put(ProducerConfig.RETRY_BACKOFF_MS_CONFIG, 1000);
+        props.put(ProducerConfig.RETRIES_CONFIG, String.valueOf(3));
+        props.put(ProducerConfig.RETRY_BACKOFF_MS_CONFIG, String.valueOf(1000));
 
         switch (type) {
         case FULL_ASYNC: {
@@ -181,7 +203,7 @@ public class KafkaHelper {
         }
         case SYNC_NO_ACK:
         default: {
-            props.put("request.required.acks", "0");
+            props.put(ProducerConfig.ACKS_CONFIG, "0");
             props.put("producer.type", "sync");
             break;
         }
@@ -283,14 +305,18 @@ public class KafkaHelper {
         props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
         props.put(ConsumerConfig.GROUP_ID_CONFIG, consumerGroupId);
         props.put(ConsumerConfig.CLIENT_ID_CONFIG, consumerGroupId);
-        props.put(ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG, 3000);
-        props.put(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG, 5000);
-        props.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, 7000);
-        props.put(ConsumerConfig.REQUEST_TIMEOUT_MS_CONFIG, 9000);
+
+        props.put(ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG, String.valueOf(1000));
+
+        props.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, String.valueOf(6000));
+        props.put(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG, String.valueOf(2000));
+
+        props.put(ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG, String.valueOf(60000));
+        props.put(ConsumerConfig.REQUEST_TIMEOUT_MS_CONFIG, String.valueOf(61000));
 
         if (autoCommitOffset) {
             props.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, true);
-            props.put(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG, 1000);
+            props.put(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG, String.valueOf(1000));
         } else {
             props.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, false);
         }
@@ -304,18 +330,18 @@ public class KafkaHelper {
 
         if (leaderAutoRebalance) {
             props.put("auto.leader.rebalance.enable", true);
-            props.put("rebalance.backoff.ms", 10000);
-            props.put("refresh.leader.backoff.ms", 1000);
+            props.put("rebalance.backoff.ms", String.valueOf(10000));
+            props.put("refresh.leader.backoff.ms", String.valueOf(1000));
         } else {
             props.put("auto.leader.rebalance.enable", false);
         }
 
         props.put("controlled.shutdown.enable", true);
-        props.put("controlled.shutdown.max.retries", 3);
-        props.put("controlled.shutdown.retry.backoff.ms", 3000);
+        props.put("controlled.shutdown.max.retries", String.valueOf(3));
+        props.put("controlled.shutdown.retry.backoff.ms", String.valueOf(3000));
 
-        // max 1mb
-        props.put(ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG, 1 * 1024 * 1024);
+        // max 256kb
+        props.put(ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG, String.valueOf(256 * 1024));
 
         if (customProps != null) {
             // populate custom configurations